Migrate DragDropContext and Droppable to FutureViewDayTaskContainer component
