+++ frontend/src/components/TaskView/FutureView/FutureViewDayContent.tsx
@@ -1,12 +1,9 @@
 import React, { ReactElement } from 'react';
-import { DragDropContext, Droppable, DropResult } from 'react-beautiful-dnd';
 import { SimpleDate } from './future-view-types';
 import { CalendarPosition, FloatingPosition } from '../../Util/TaskEditors/editors-types';
 import styles from './FutureViewDay.module.scss';
 import { day2String, getTodayAtZeroAM } from '../../../util/datetime-util';
 import FutureViewDayTaskContainer from './FutureViewDayTaskContainer';
-import { computeReorderMap, getReorderedList } from 'util/order-util';
-import { applyReorder } from 'firebase/actions';
 
 type Props = {
   readonly date: SimpleDate;
@@ -17,8 +14,6 @@ type Props = {
   readonly inMainList: boolean;
   readonly onHeightChange: (doesOverflow: boolean, tasksHeight: number) => void;
 };
-
-type IdOrder = { readonly id: string; readonly order: number };
 
 /**
  * The main content of future view day.
@@ -36,48 +31,23 @@ function FutureViewDayContent(
 ): ReactElement {
   const containerStyle = (inNDaysView && inMainList) ? { paddingTop: '1em' } : {};
   const isToday: boolean = getTodayAtZeroAM().toDateString() === date.text;
-  const onDragEnd = (result: DropResult): void => {
-    const { source, destination } = result;
-    if (destination == null) {
-      // invalid drop, skip
-      return;
-    }
-    let sourceOrder: number;
-    let destinationOrder: number;
-    sourceOrder = source.index;
-    const dest = destination.index;
-    destinationOrder = dest == null ? sourceOrder : 0;
-
-    // const reorderMap = computeReorderMap(localTasks, sourceOrder, destinationOrder);
-    // setLocalTasks(getReorderedList(localTasks, reorderMap));
-    // applyReorder('tasks', reorderMap);
-  };
   return (
     <>
-      <DragDropContext onDragEnd={onDragEnd}>
-        <div className={styles.DateInfo} style={containerStyle}>
-          <div className={styles.DateInfoDay}>
-            {isToday ? 'TODAY' : day2String(date.day)}
-          </div>
-          <div className={styles.DateNum}>{date.date}</div>
+      <div className={styles.DateInfo} style={containerStyle}>
+        <div className={styles.DateInfoDay}>
+          {isToday ? 'TODAY' : day2String(date.day)}
         </div>
-        <Droppable droppableId={`${date.month}/${date.day}/${date.year}`}>
-          {(provided) => (
-            <div ref={provided.innerRef} {...provided.droppableProps} {...provided.droppableProps}>
-              <FutureViewDayTaskContainer
-                date={date.text}
-                inNDaysView={inNDaysView}
-                taskEditorPosition={taskEditorPosition}
-                calendarPosition={calendarPosition}
-                doesShowCompletedTasks={doesShowCompletedTasks}
-                isInMainList={inMainList}
-                onHeightChange={onHeightChange}
-              />
-              {provided.placeholder}
-            </div>
-          )}
-        </Droppable>
-      </DragDropContext>
+        <div className={styles.DateNum}>{date.date}</div>
+      </div>
+      <FutureViewDayTaskContainer
+        date={date.text}
+        inNDaysView={inNDaysView}
+        taskEditorPosition={taskEditorPosition}
+        calendarPosition={calendarPosition}
+        doesShowCompletedTasks={doesShowCompletedTasks}
+        isInMainList={inMainList}
+        onHeightChange={onHeightChange}
+      />
     </>
   );
 }
